pyproject.toml
[project]
name = "subagent-poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.1",
    "langchain-openai>=1.1.3",
    "openai>=2.12.0",
    "psycopg2[binary,pool]>=2.9.11",
    "python-dotenv>=1.2.1",
    "streamlit>=1.52.1",
]
